refactor: move ChatMessage to components
